Moved datatools to its own class
